+++ database/migrations/2020_01_01_000003_create_plan_subscriptions_table.php
@@ -15,9 +15,9 @@ class CreatePlanSubscriptionsTable extends Migration
     public function up(): void
     {
         Schema::create(config('rinvex.subscriptions.tables.plan_subscriptions'), function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->morphs('user');
-            $table->integer('plan_id')->unsigned();
+            $table->bigInteger('plan_id')->unsigned();
             $table->string('slug');
             $table->{$this->jsonable()}('name');
             $table->{$this->jsonable()}('description')->nullable();
